+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New #186: Add `ChatOwnerLeft` and `ChatOwnerChanged` types.
 - New #186: Add `chatOwnerLeft` and `chatOwnerChanged` fields to `Message` type.
 - New #186: Add `SetMyProfilePhoto` method.
+- New #186: Add `RemoveMyProfilePhoto` method.
 
 ## 0.14 February 7, 2026